move types into type file
