+++ src/components/views/settings/devices/filter.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { DeviceWithVerification } from "./useOwnDevices";
+import { DeviceWithVerification } from "./types";
 
 export enum DeviceSecurityVariation {
     Verified = 'Verified',
